:truck: moves sections into components folder
